fix: complete stage 1 strangler guardrails
