+++ .github/workflows/ci.yml
@@ -54,7 +54,7 @@ jobs:
       run: |
         # Ruff has no max-module-lines rule. This check prevents new files from
         # exceeding the current worst case. Tighten the threshold over time.
-        MAX_LINES=2450   # current max: 2404 (github_downloader.py)
+        MAX_LINES=1400   # Stage 1 strangler target (#1077)
         VIOLATIONS=$(find src/ -name '*.py' -print0 | xargs -0 -I{} awk -v max="$MAX_LINES" \
           'END { if (NR > max) printf "%s: %d lines (max %d)\n", FILENAME, NR, max }' {})
         if [ -n "$VIOLATIONS" ]; then
